--- a/src/set.jl
+++ b/src/set.jl
@@ -16,8 +16,8 @@ function AVLSet(x::K) where {K <: AbstractVector}
 end
 
 Base.eltype(::Type{AVLSet{K}}) where {K} = K
-Base.length(set::AVLSet) = length(set.tree) 
-Base.in(x::K, set::AVLSet{K}) where {K} = find_node(set.tree, x) !== nothing 
+Base.length(set::AVLSet) = length(set.tree)
+Base.in(x::K, set::AVLSet{K}) where {K} = x in set.tree
 
 function iterate(set::AVLSet{K}) where {K}
     ret = iterate(set.tree)