Things checked out:

- replace `isnothing` 6s -> 5.18s
- add additional dispatch in rotations 5.18s -> 3.7s
- barriers on rebalance, more type stability on insert: now at 2.52s
- more things: now at 1.87s


benchmark
```
using BenchmarkTools, AVLTrees
m=1000
n=1000

function testbase5(m, n)
    tree = AVLTrees.AVLSet{Int}()
    for i in 1:m
        for j in 1:n
            push!(tree, j)
            push!(tree, -j)
        end
        @assert !(0 in tree)
        for j in 1:n
            @assert j in tree
            @assert -j in tree
        end
        @assert !(n + 1 in tree)
        @assert !(-n - 1 in tree)
        for j in 1:n
            @assert j in tree
            @assert -j in tree
            delete!(tree, j)
            delete!(tree, -j)
            @assert !(j in tree)
            @assert !(-j in tree)
        end
    end
end

@btime testbase5(m,n)
```